Fork test fixes post release (#156)
After running the last Baklava release, we noticed a limitation of the fork tests where bucket sizes were not taken into account when doing trades towards the limits. This PR fixes this by changing the swap amount logic to include buckets and also simulating oracle updates so that the buckets rebalance.