Airgrab Contract (#274)
After a few iterations, we landed on a simple Airgrab scheme:
- With a list of <address>,<token amount> we generate a merkle tree and
record the root in the contract
- A user does KYC, gets the merkle proof from our backend and submits
the claim
- The contract verifies the merkle proof and KYC and locks the tokens on
behalf of the user for a predetermined cliff and slope.
- The airgrab has an endTimestamp after witch any leftover tokens in the
contract can be transferred to the treasury.

More details on specifics around the KYC can be extracted from the
comments in the code which contain links to documentation.

There is a single open question around how we encode the expected
credential from fractal. The expected credential is basically the set of
restrictions we want to enforce and it's a string that looks like
`"level:plus;residency_not:ca,us"`.
I currently provided that statically in the code and it's probably not
the final version that we want. But to parameterize in a naive way would
incur a big gas because we can only store it in storage, and not as an
immutable parameter because only immutable base types are allowed as of
now in solidity. If we want to parameterize it the approach I would take
would be to store a keccak of it as an immutable bytes32 parameter, and
provide the full string as a parameter to the claim function and when
verifying the credential we would first verify that the keccak matches
and then use it in the construction of the full credential. I'm not sure
if it's worth it or if it's fine to just come up with the value and keep
it static in line as is right now.

N/A

Yes

- Fixes #282

N/A

N/A

---------

Co-authored-by: baroooo <baranseltekin@gmail.com>