Split mento allocation (#286)
### Description

As discussed internally, we want to split the vesting allocation into
two.
1- First part will be sent to vesting as initially planned (8%)
2- Second part will be sent to a multisig with a timelock and will be
ready to be used by Mento Labs for various purposes, but the governance
will have veto rights (12%)

This ticket is only about the allocation of the funds. Multisig will be
added later.

### Other changes

Airgrab contract was using rarible's locking repo as module to use the
locking interface. Since the locking-contract changes are merged, we use
the local interface instead and removed the module. See the first
resolved conversation here:

https://github.com/mento-protocol/mento-core/pull/274/files#diff-fe7afb5c9c916e521401d3fcfb4277d5071798c3baf83baf11d6071742823584R16

### Tested

Updated unit tests

### Related issues

- Fixes #285

---------

Co-authored-by: Bayological <6872903+bayological@users.noreply.github.com>
Co-authored-by: philbow61 <philip.raetsch@gmail.com>
Co-authored-by: boqdan <304771+bowd@users.noreply.github.com>