feat: make Token pausable (#366)
Based on recent token-related conversations we've decided we want more
flexibility when it comes to the time between when we launch governance
and give utility to the token, and the time when the token is
transferable, i.e. tradable.
Most recent token launches have followed this approach, most notably
Safe.
These changes make the token ownable and pausable. 

However, even when paused the token can still be transferred by:
- The owner
- The locking contract - it can be locked and unlocked, thus the airgrab
can run while paused
- The emissions contract - it can be minted into existence while paused