feat: make ci ✅ with 💚 and 🥑s (#371)
### Description

- Split jobs for better readability
- Make storage checks run only on push to develop (i.e. not for each
commit of a work in progress branch) in an attempt to not hit rate
limits.
- removed the slither.db.json and inlined ignores - it's important to
actually review these and see if we want to actually fix any of them
- made slither ignore contracts/common and contracts/legacy . Common
will be moved to @celo/contracts npm dependency anyway.
- fixed echidna compilation issue, yey!

### Other changes

- Moved `SortedOracles.sol` to `common` because moving forward that
contract will be owned by `clabs` and we will need to deploy it through
their pipeline. So everything in `common` will move to
`@celo/contracts`, including SortedOracles.