feat: implement Chainlink relayer contract (#456)
### Description

Adds a ChainlinkAdapter contract which can relay anwers from a Chainlink
AggregatorV3 to SortedOracles.

It assumes it is the only reporter for the given rate feed in
SortedOracles (e.g. `greater`/`lesserKey` are not needed).

Every rate feed that relies on this mechanism should have a separate
ChainlinkAdapter deployed, with the appropriate parameters passed to the
constructor. To save gas, the adapter is stateless, unupgradeable, and
relies on immutable parameters set in the constructor.

Next steps:

* Integration test of the contract.
* Off-chain component that can ping this adapter contract.
* Deployment tooling/docs for setting up a rate feed with the
ChainlinkAdapter.

### Tested

Unit tests.

### Related issues

- Fixes #458

---------

Co-authored-by: chapati <philip.paetz@me.com>