Feat: Make ChainlinkRelayer support implicit pairs (#487)
### Description

Late last week it became apparent that Chainlink won't provide all
derived rates that we need, and that we will need to compute implicit
rates on-chain. This PR extends the relayer such that it can aggregate
up to 4 chainlink date feeds into a single price. Each data feed can be
inverted (1 / p) to make the dominos line up.

### Other changes

Upgraded the tests to solidity 0.8 and fixed the issue with deploying
SortedOracles via the factory by pre-deploying libraries as well because
then factory links them automatically when calling `vm.getCode`.

### Tested

Tests have been extended, and I've also used a slightly new pattern that
I didn't use before which I like, and might not be evident from reading
the code. Specifically for the relayer we have 4 scenarios depending on
how many price feeds it aggregates, so I wrote test contracts that
extend each other in such a way where test cases from N-1 price feeds
get executed also in the test for N price feeds.

### Related issues

N/A

### Backwards compatibility

Not backwards compatible but not deployed yet.

### Documentation

N/A

---------

Co-authored-by: Marcin Chrzanowski <martin@clabs.co>
Co-authored-by: Martin <marcin.j.chrzanowski@gmail.com>
Co-authored-by: chapati <philip.paetz@me.com>
Co-authored-by: Nelson Taveras <4562733+nvtaveras@users.noreply.github.com>
Co-authored-by: baroooo <baranseltekin@gmail.com>