feat/cosmetic: Move Proxy Contracts around (#498)
### Description

When we restructured the repo a while ago I moved the proxy contracts to
the legacy folder thinking that we might not deploy named proxies in the
future, and maybe upgrade to a more modern Proxy implementation, but we
have continued to add the proxies inside the `legacy` folder which
doesn't feel great as that folder is meant to be removed (and probably
can be at this point).

So in order to embrace what we're doing I propose collocating the proxy
contracts to the `contracts/tokens` directory. And as an addition to
that, I propose collocating all proxies with their contracts like Martin
also did in the `contracts/oracles` for the new ChainlinkRelayerFactory
contract and proxy. So I also moved around the proxies in
`contracts/proxies`.

This shouldn't really effect anything else.

### Other changes

- Renamed stabletokenPSO to stabletokenPHP

### Tested

Ran tests, all is well.

### Related issues

N/A

### Backwards compatibility

YES

### Documentation

N/A

---------

Co-authored-by: Bayological <6872903+bayological@users.noreply.github.com>
Co-authored-by: denviljclarke <60730266+denviljclarke@users.noreply.github.com>