Remove the need for `--via-ir` in our contracts (#502)
### Description

After getting more and more annoyed with the CI time and other small
issues around this I decided to really understand how big is our need
for `--via-ir` compilation. Turns out it's really small. I mean really
small.

There are just a couple of places where this is causing issues: 

1. `Airgrab#claim` has too many variables. I fixed this with a `struct`
for the fractal proof, and an internal function to do the lock. Changing
it is ok in my opinion. If we want to reuse this contract we'll have to
redeploy it anyway, we have a tag for the version when it was deployed,
all is well.

2. `GovernanceFactory#createGovernance` has too many variables. I just
broke it down in internal functions. Again, I don't see a problem in
changing this, even if it's deployed and etc.

3. `StableTokenV2#initialize` has too many arguments. When we made
StableTokenV2 we wanted to keep the same interface for the `initialize`
method to keep the interfaces between stable tokens consistent, but it
was actually a bad call. There's no good reason why we have it that way
and it's a method that can anyway only be called once. I would say it's
ok to also change this without making a StableTokenV3 for it. Basically
we will have a few contracts which will have a different signature for
the initialize method that anyway can't be called anymore. But this is
the one change the feels least nice.

4. A couple of tests with many variables where I used structs for
grouping or broke down into internal functions.

What happens afterwards, well:
```
╰─ forge test
[⠊] Compiling...
[⠒] Compiling 64 files with Solc 0.5.17
[⠢] Compiling 55 files with Solc 0.8.26
[⠒] Compiling 223 files with Solc 0.8.18
[⠒] Solc 0.5.17 finished in 307.90ms
[⠘] Solc 0.8.26 finished in 1.75s
[⠒] Solc 0.8.18 finished in 5.50s
Compiler run successful with warnings:
```
The whole fucking shabang runs in 5.5s. Jesus.
CI takes ~1minute instead of ~15minutes.

P.S. Thank you Claude for the bash script.

### Other changes

It changed my life.

### Tested

Thoroughly.

### Related issues

Sleep deprivation.

### Backwards compatibility

Who are you calling backward?

### Documentation

I read it, yes.

---------

Co-authored-by: Bayological <6872903+bayological@users.noreply.github.com>
Co-authored-by: chapati <philip.paetz@me.com>
Co-authored-by: philbow61 <80156619+philbow61@users.noreply.github.com>
Co-authored-by: baroooo <baranseltekin@gmail.com>