fix: gas tests (#507)
### Description

Context: We create 50k locks to check how much gas is used by governance
and locking operations for users with high number of locks. It was
working fine until a couple of weeks ago. Then we disabled the gas tests
to make CI happy.

The problem was related to a [recent change on
foundry](https://github.com/foundry-rs/foundry/pull/8274/files). They
reduced the default gas limit for tests from `9223372036854775807` to
`1073741824` which is not enough to create as many locks as we want.

This PR fixes the issue by setting the gas limit to the previous default
value of `9223372036854775807` or `i64::MAX / 2` and re-enables the gas
tests. I don't think this will cause any problems as the only downside
of this is if we have an infinite loop in our tests, it will take some
time to consume all the available gas before the test fails.


### Other changes

- Used a common var `viewCallGas` instead of creating one for each test
- Increased the expected view call gas usages from 20k to 30k as they
consume slightly higher gas now, probably thats because we removed the
gas optimization across repo recently

### Tested

✅ 

### Related issues

- Fixes
[#495](https://github.com/mento-protocol/mento-general/issues/495)