Feat/fpmm (#595)
### Description

This PR introduces the Fixed Price Market Maker contract, a core
component of the Mento v3 protocol that facilitates token swaps at fixed
exchange rates. The FPMM manages a liquidity pool of two tokens and
enables trading between them based on oracle rates. It implements a
market maker model similar to Uniswap and uses an interface similar to
Velodrome but with pricing based on external oracles and with a
flash-loan-like rebalancing mechanism for various liquidity strategies.

Key features of the FPMM implementation:
- Oracle-based pricing for token swaps using the SortedOracles contract.
This is subject to change and we might swap SorterOracles with another
implementation.
- Support for rebalancing mechanism to realign pool reserves with oracle
prices
- Circuit breaker integration via BreakerBox contract
- Liquidity provision mechanism with ERC20 LP tokens
- Fee management with configurable protocol fees
- Decimal-aware calculations supporting tokens with different decimal
places

### Tested
The implementation is covered by unit and integration tests.
Fuzz tests will be added in a separate PR

---------

Co-authored-by: philbow61 <philip.raetsch@gmail.com>