Feat: router (#610)
This PR adds a Uniswapv2 like router which is deployed here by
Aerodrome:
https://vscode.blockscan.com/8453/0xcF77a3Ba9A5CA399B7c97c74d54e5b1Beb874E43
Added openzeppelin contracts inherited by the Router manually to ensure
the exact same code is used.
Tried to change as less as possible from the Router. This is why we did
not touch things like WETH handling even tho it does not apply to Celo.
Keeping WETH address at 0 should be good enough to block those calls

- Ensures fpmms are deployed with token addresses in sorted order
- Adds some helpers to FPMMFactory
- Removes Liquityv2 Contract and other related changes previously merged
because we decided to keep them in another repo
- Adds integration tests between Factory <-> FPMM <-> Router. Other
components are not implemented yet so tests related to those are
supposed to be added later
- Fixes slither issues
- Not sure what happened to Factory unit tests on this branch but I just
re-added those

- Ensure the core router logic is not altered
- We did not introduce any artifacts due to all the branch juggling
- FPMMFactory should work as before
- Token ordering is implemented correctly