use `extcodehash` for contract codehash check (#619)
### Description

To ensure we have the right `CreateX` contract, we perform a check on
its codehash. Previously, this was done by loading the entire code at
the target address into memory and hashing it, this update makes use of
the `extcodehash` opcode which allows us to make huge gas savings (4357
gas units in this case).

### Tests & Backwards compatibility

Absolutely no side effect is expected.