V3 Router / Broker adapter (#617)
### Description

This PR introduces `VirtualPool` contracts and their factory. These are
contracts that are compatible with the new Router but actually just run
swaps through the Broker.

### Other changes

- Moved some methods from `IFPMM` to `IRPool`
- Removed `registry.setAddressFor` in tests to allow simpler fork tests
without pranking the registry owner
- Added a simple multi-hop swap in Router tests
- Fixed some typos in comments

### Things that still need a discussion

We need to unify how we deploy. It can be CREATE2 or clone or a mix of
CREATE2 and clone, or it can remain CreateX