Adopt BUSL License for v3 (#663)
### Description
Similarly to many other large DeFi players, we're adopting the BUSL
License. Our goal is to achieve a good trade-off between honoring the
open source spirit, that built the industry we love, while at the same
time protecting us from low effort hostile forks.

And after studying other industry-leading protocols, we came to the
conclusion that BUSL is the best trade-off for us.

### To Do
- [x] Update LICENSE file
- [x] Update SPDX statements in Solidity files

### Prior art
- Morpho: https://github.com/morpho-org/morpho-blue/blob/main/LICENSE
- Uniswap v4:
https://support.uniswap.org/hc/en-us/articles/33829751588109-Uniswap-v4-licensing
- Uniswap v3: https://github.com/Uniswap/v3-core/blob/main/LICENSE
- Aave: https://github.com/aave-dao/aave-v3-origin/blob/main/LICENSE
- Liquity:
https://www.liquity.org/blog/licensing-liquity-v2-may-the-fork-be-with-you

---------

Co-authored-by: Nelson Taveras <4562733+nvtaveras@users.noreply.github.com>